CHORE: rename _io.mrat to _io.stc_dataset
