+++ eelbrain/_io/stc_dataset.py
